1st Assignment-KIM
Made a new page by using example1.
@fraquada, @pearswj
I tried to move Javascript part to script.jb at the new window of Javascript. But it doesn't work.
So that like exemple1, I mix CSS and JS on the HTML page. I'd love to know how to seperate.

2nd problem : I'd like to give different image size for each image.
How can we do it?